Adapt to Kotlin target level 2.3
Most of these changes are related to improved type inference for
null-annotated Java APIs, especially Guava container classes. In some
cases this exposed real issues in prod, now fixed.

There were also a handful of removed kotlin-stdlib APIs.

Bug: 446228176
Test: existing

+++ fakeadbserver/src/main/java/com/android/fakeadbserver/shellcommandhandlers/ShellProtocolEchoCommandHandler.kt
@@ -87,7 +87,7 @@ class ShellProtocolEchoCommandHandler() : SimpleShellHandler(ShellProtocolType.S
     fun process(bytes: ByteArray) {
       for (b in bytes) {
         buffer.put(b)
-        if (b.toChar() == '\n') {
+        if (b.toInt().toChar() == '\n') {
           flush()
         }
       }
